move PureKerasModel and get_input_vector into shared module
